+++ system-tests/projects/create-react-app-configured/src/App.cy.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import App from './App'
+
+describe('<App />', () => {
+  it('renders', () => {
+    // see: https://on.cypress.io/mounting-react
+    cy.mount(<App />)
+    cy.contains('Learn React')
+  })
+})