+++ system-tests/projects/create-react-app-configured/src/App.cy.jsx
@@ -1,8 +1,0 @@
-import React from 'react'
-import { mount } from 'cypress/react'
-import App from './App'
-
-it('works', () => {
-  mount(<App />)
-  cy.contains('Learn React')
-})